Major cleanup and reorganization of the plugin
This maintains the same functionality as before but fixes some minor issues and
adds options and comments to the code. The major changes is that you can now
kill or switch to views that are not in the active group. You can also auto kill
a specific index in the list (0 for the current view). This can replace the
standard super+w shortcut which so that you jump to the most recently used tab
after closing.