Fix an issue where the kill command changed the order of active buffers
When closing a buffer sublime automatically switches to a different view. This
plugin changes that behavior to switch to a view based on which was most
recently used. While doing this correction though I was leaving the code on that
tracks switches which created a bunch of phantom activity in the recently used
list. Now we will ignore all switches that happen when killing a view.